+++ src/index.ts
@@ -3,3 +3,4 @@ export * from './Wizard';
 export * from './adapters/AdapterInterface';
 export * from './adapters/DiscordJSAdapter';
 export * from './adapters/ErisAdapter';
+export * from './validators';